Reverted to use wrapper only
